% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citation}
\alias{citation}
\title{Get citation function}
\usage{
citation(variables)
}
\arguments{
\item{variables}{variable names from cspp, as strings}
}
\description{
This function will pull the citations for a variable(s)
    in the Correlates of State Policy Project data.
}
