% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{The Correlates of State Policy Project Dataset Codebook}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2090 rows and 12 columns.
}
\usage{
codebook
}
\description{
@docType data
}
\details{
@usage data(codebook)

 @format The Correlates of State Policy Project dataset codebook. Each observation is a variable in the CSPP dataset. There are 2090 rows and 12 columns. The columns contain information regarding each variable, such as its source, citation, and a short & long description.

 @keywords datasets
}
\keyword{datasets}
