% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd2HTML.R
\name{convert}
\alias{convert}
\title{Convert Rd files to HTML for documentation}
\usage{
convert(RdNames)
}
\arguments{
\item{Rdname}{name of RD file}

\item{labels}{data frame of variable labels,
columns are variable, obs are descriptions}
}
\description{
This function takes Rd output files (documentation files
    that R uses), and converts them to HTML for documentation
    that fits course requirement.
}
