% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlates.R
\docType{data}
\name{correlates}
\alias{correlates}
\title{The Correlates of State Policy Project Dataset (IPPSR)}
\format{
An object of class \code{data.frame} with 6120 rows and 2091 columns.
}
\usage{
correlates
}
\description{
The Correlates of State Policy Project dataset assembled by Marty P. Jordan and Matt Grossmann (2020) <http://ippsr.msu.edu/public-policy/correlates-state-policy>. Each observation is a state-year.
}
\details{
@name correlates

 @docType data

 @usage data(correlates)


 @keywords datasets
}
\keyword{datasets}
