% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullDescribe.R
\name{fullDescribe}
\alias{fullDescribe}
\title{Get full description function}
\usage{
fullDescribe(variables)
}
\arguments{
\item{variables}{variable names from cspp, as strings}
}
\description{
This function will pull the long description for a variable(s)
    in the Correlates of State Policy Project data.
}
