% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_example.R
\docType{data}
\name{map_example}
\alias{map_example}
\title{Sample Dataset for Working with generate_map()}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
}
\usage{
map_example
}
\description{
A dataframe to create a sample map using the generate_map function. The variable plotted is population.
}
\details{
@name map_example

 @docType data

 @usage data(map_example)

 @keywords datasets
}
\keyword{datasets}
