% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{State Network data (IPPSR)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2550 rows and 120 columns.
}
\usage{
network_data
}
\description{
@docType data
}
\details{
@usage data(network_data)

 @format A data frame of n rows and n variables. Each observation is ...

 @keywords datasets
}
\keyword{datasets}
