% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_vars.R
\docType{data}
\name{network_vars}
\alias{network_vars}
\title{Variables in the State Network (IPPSR) data}
\format{
An object of class \code{data.frame} with 118 rows and 2 columns.
}
\usage{
network_vars
}
\description{
@name network_vars
}
\details{
@docType data

 @usage data(network_vars)

 @format A dataset of the the names of the variables in the IPPSR state networks data.

 @keywords datasets
}
\keyword{datasets}
