% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickDescribe.R
\name{quickDescribe}
\alias{quickDescribe}
\title{Get quick description function}
\usage{
quickDescribe(variables)
}
\arguments{
\item{variables}{variable names from cspp, as strings}
}
\description{
This function will pull the short description for a variable(s)
    in the Correlates of State Policy Project data.
}
