% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_names_db.R
\docType{data}
\name{var_names_db}
\alias{var_names_db}
\title{Variables Names of the Correlates of State Policy Project Dataset (IPPSR)}
\format{
An object of class \code{data.frame} with 2179 rows and 3 columns.
}
\usage{
var_names_db
}
\description{
@name var_names_db
}
\details{
@docType data

 @usage data(var_names_db)

 @format A dataframe of all variable names, their descriptions, and sources in the Correlates of State Policy Project Dataset.

 @keywords datasets
}
\keyword{datasets}
