% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{years}
\alias{years}
\title{Get years function}
\usage{
years(variables)
}
\arguments{
\item{variables}{variable names from cspp, as strings}
}
\description{
This function will print the years available for a variable(s)
    in the Correlates of State Policy Project data.
}
