% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{The Correlates of State Policy Project Dataset Codebook}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2090 rows and 12 columns.
}
\usage{
codebook
}
\description{
@name codebook
}
\details{
@docType data

 @usage data(codebook)

 @format The codebook for the Correlates of State Policy Project dataset assembled by Marty P. Jordan and Matt Grossmann (2020) <http://ippsr.msu.edu/public-policy/correlates-state-policy>. Each observation is a variable in the CSPP dataset. The columns contain information regarding each variable, such as its source, citation, and a short & long description.

 @keywords datasets
}
\keyword{datasets}
