% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Codebook for CPSS data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2090 rows and 12 columns.
}
\usage{
codebook
}
\description{
@docType data
}
\details{
@usage data(codebook)

 @format A data frame of n rows and n variables. Each observation is ...

 @keywords datasets
}
\keyword{datasets}
