% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlates.R
\docType{data}
\name{correlates}
\alias{correlates}
\title{correlates}
\format{
An object of class \code{data.frame} with 6120 rows and 2091 columns.
}
\usage{
correlates
}
\description{
@docType data
}
\details{
@usage data(correlates)

 @format A data frame of n rows and n variables. Each observation is ...

 @keywords datasets
}
\keyword{datasets}
