% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_example.R
\docType{data}
\name{map_example}
\alias{map_example}
\title{map_example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
}
\usage{
map_example
}
\description{
@docType data
}
\details{
@usage data(map_example)

 @format A data frame of n rows and n variables. Each observation is ...

 @keywords datasets
}
\keyword{datasets}
