% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{State Network data (IPPSR)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2550 rows and 120 columns.
}
\usage{
network_data
}
\description{
The State Networks dataset is a compilation of many state-to-state relational variables, including measures of shared borders, travel and trade between states, and demographic characteristics of state populations collected by Shayla Olson (2020) and Marty P. Jordan and Matt Grossmann (2020) <http://ippsr.msu.edu/public-policy/state-networks>.
}
\details{
@name network_data

 @docType data

 @usage data(network_data)

 @keywords datasets
}
\keyword{datasets}
