% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_vars.R
\docType{data}
\name{network_vars}
\alias{network_vars}
\title{Variables in the State Network (IPPSR) data}
\format{
An object of class \code{data.frame} with 118 rows and 2 columns.
}
\usage{
network_vars
}
\description{
@docType data
}
\details{
@usage data(network_vars)

 @format A data frame of n rows and n variables. Each observation is ...

 @keywords datasets
}
\keyword{datasets}
