% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_vars.R
\docType{data}
\name{network_vars}
\alias{network_vars}
\title{State Network (IPPSR) Dataset Variable Names}
\format{
An object of class \code{data.frame} with 118 rows and 2 columns.
}
\usage{
network_vars
}
\description{
A dataset of the the names of the variables in the IPPSR state networks data.
}
\details{
@name network_vars

 @docType data

 @usage data(network_vars)

 @keywords datasets
}
\keyword{datasets}
