% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_names_db.R
\docType{data}
\name{var_names_db}
\alias{var_names_db}
\title{var_names_db}
\format{
An object of class \code{data.frame} with 2179 rows and 3 columns.
}
\usage{
var_names_db
}
\description{
@docType data
}
\details{
@usage data(var_names_db)

 @format A data frame of n rows and n variables. Each observation is ...

 @keywords datasets
}
\keyword{datasets}
